Refactored vuex store
